refactor ConfigValidator
