+++ src/ReverseProxy/Management/IReverseProxyBuilderExtensions.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Yarp.ReverseProxy.Configuration;
+using Yarp.ReverseProxy.Configuration.ClusterValidators;
+using Yarp.ReverseProxy.Configuration.RouteValidators;
 using Yarp.ReverseProxy.Delegation;
 using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Health;
@@ -25,6 +27,17 @@ internal static class IReverseProxyBuilderExtensions
     {
         builder.Services.TryAddSingleton<IYarpRateLimiterPolicyProvider, YarpRateLimiterPolicyProvider>();
         builder.Services.TryAddSingleton<IConfigValidator, ConfigValidator>();
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IRouteValidator, HeadersValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IRouteValidator, HostValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IRouteValidator, MethodsValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IRouteValidator, PathValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IRouteValidator, QueryParametersValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, DestinationValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, LoadBalancingValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, HealthCheckValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, SessionAffinityValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, ProxyHttpClientValidator>());
+        builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IClusterValidator, ProxyHttpRequestValidator>());
         builder.Services.TryAddSingleton<IRandomFactory, RandomFactory>();
         builder.AddTransformFactory<ForwardedTransformFactory>();
         builder.AddTransformFactory<HttpMethodTransformFactory>();
